:recycle: Types module initiated and refactored across available methods in v2
