+++ src/v2/public_api/latest_buy_price_market.rs
@@ -1,7 +1,14 @@
 use reqwest::StatusCode;
 use serde::{Deserialize, Serialize};
 
-use crate::v2::{CoinSpotBadResponse, CoinSpotPublic, CoinSpotResponse, CoinSpotResult};
+use crate::v2::{
+    CoinSpotPublic,
+    types::{
+        CoinSpotBadResponse,
+        CoinSpotResponse,
+        CoinSpotResult,
+    }
+};
 
 #[derive(Serialize, Deserialize, Debug)]
 pub struct LatestBuyPrice {
